# backend/api/games/oddoneout/openapi.yaml
openapi: 3.1.0
info:
  title: Odd One Out Game Server API
  version: 0.1.0
servers:
  - url: http://localhost:8081
paths:
...
  /v1/admin/stats:
    get:
      operationId: getOddOneOutAdminStats
      summary: Return operational metadata for coordinator aggregation
      security:
        - adminBearerAuth: []
      responses:
        "200":
          description: Current server statistics and question-pack metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameServerStats"
        "401":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/rooms:
    post:
      operationId: createRoomOnGameServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Room and host session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/rooms/search:
    get:
      operationId: searchJoinableOddOneOutRooms
      summary: Search lobby rooms that still have player capacity
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 60
      responses:
        "200":
          description: Joinable Odd One Out rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomSearchResults"
        "400":
          $ref: "#/components/responses/Problem"
  /v1/rooms/{roomId}:
    get:
      operationId: getRoomState
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Player-specific room state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomView"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/rooms/{roomId}/players:
    post:
      operationId: joinRoom
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [display_name, password]
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 30
                password:
                  type: string
                  maxLength: 100
      responses:
        "201":
          description: Player session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/rooms/{roomId}/ws:
    get:
      operationId: connectToRoom
      summary: Upgrade to the Skewa room WebSocket protocol
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "101":
          description: WebSocket connection established
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminBearerAuth:
      type: http
      scheme: bearer
      description: Long administrator password configured through ADMIN_PASSWORD
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateRoomRequest:
      type: object
      additionalProperties: false
      required: [name, password, host_name, settings]
      oneOf:
        - required: [question_pack]
        - required: [questions]
      properties:
        room_id:
          type: string
          description: Optional coordinator-assigned identifier
          pattern: "^[A-Za-z0-9_-]{1,100}$"
        name:
          type: string
          minLength: 1
          maxLength: 60
        password:
          type: string
          maxLength: 100
        host_name:
          type: string
          minLength: 1
          maxLength: 30
        settings:
          $ref: "#/components/schemas/Settings"
        question_pack:
          type: string
          description: Identifier of a built-in backend-owned question pack
        questions:
          type: array
          writeOnly: true
          description: Custom question pairs supplied as room-creation input
          minItems: 1
          items:
            $ref: "#/components/schemas/QuestionInput"
    Settings:
      type: object
      additionalProperties: false
      required: [player_limit, answer_seconds, discussion_seconds, voting_seconds, rounds]
      properties:
        player_limit:
          type: integer
          minimum: 3
          maximum: 20
        answer_seconds:
          type: integer
          minimum: 5
          maximum: 900
        discussion_seconds:
          type: integer
          minimum: 5
          maximum: 3600
        voting_seconds:
          type: integer
          minimum: 5
          maximum: 900
        rounds:
          type: integer
          minimum: 1
          maximum: 50
    QuestionInput:
      type: object
      additionalProperties: false
      required: [real, fake]
      description: Write-only custom question input; never returned in room state
      properties:
        real:
          type: string
          minLength: 1
          maxLength: 500
        fake:
          type: string
          minLength: 1
          maxLength: 500
...
    GameServerStats:
      type: object
      additionalProperties: false
      required: [game, rooms, players, question_packs]
      properties:
        game:
          type: object
          additionalProperties: false
          required: [id, name]
          properties:
            id: { const: oddoneout }
            name: { const: Odd One Out }
        rooms:
          $ref: "#/components/schemas/AdminRooms"
        players:
          $ref: "#/components/schemas/AdminPlayers"
        question_packs:
          type: array
          items:
            $ref: "#/components/schemas/AdminQuestionPack"
    AdminQuestionPack:
      type: object
      additionalProperties: false
      required: [id, name, description, question_count, items]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        question_count: { type: integer, minimum: 1 }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AdminPackItem"
    AdminPackItem:
      type: object
      additionalProperties: false
      required: [fields]
      properties:
        fields:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AdminContentField"
    AdminContentField:
      type: object
      additionalProperties: false
      required: [label, value]
      properties:
        label: { type: string, minLength: 1 }
        value: { type: string, minLength: 1 }
    AdminRooms:
      type: object
      additionalProperties: false
      required: [total, active, finished, by_phase]
      properties:
        total: { type: integer, minimum: 0 }
        active: { type: integer, minimum: 0 }
        finished: { type: integer, minimum: 0 }
        by_phase:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
    AdminPlayers:
      type: object
      additionalProperties: false
      required: [total, connected]
      properties:
        total: { type: integer, minimum: 0 }
        connected: { type: integer, minimum: 0 }
    Credentials:
      type: object
      required: [room_id, player_id, token]
      properties:
        room_id:
          type: string
        player_id:
          type: string
        token:
          type: string
    Session:
      allOf:
        - $ref: "#/components/schemas/Credentials"
        - type: object
          required: [websocket_path]
          properties:
            websocket_path:
              type: string
    RoomSearchResults:
      type: object
      additionalProperties: false
      required: [rooms]
      properties:
        rooms:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/RoomSuggestion"
    RoomSuggestion:
      type: object
      additionalProperties: false
      required: [name, game_id, game_name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        game_id:
          const: oddoneout
        game_name:
          const: Odd One Out
    Player:
      type: object
      required: [id, display_name, is_host, connected, score]
      properties:
        id:
          type: string
        display_name:
          type: string
        is_host:
          type: boolean
        connected:
          type: boolean
        score:
          type: integer
    Answer:
      type: object
      required: [player_id, player_name, text]
      properties:
        player_id:
          type: string
        player_name:
          type: string
        text:
          type: string
    RoundResult:
      type: object
      required: [imposter_id, found, vote_counts]
      properties:
        imposter_id:
          type: string
        found:
          type: boolean
        vote_counts:
          type: object
          additionalProperties:
            type: integer
    RoomView:
      type: object
      required: [version, room_id, room_name, phase, settings, players, round, your_player_id, answer_locked, vote_locked]
      properties:
        version:
          type: integer
          minimum: 0
        room_id:
          type: string
        room_name:
          type: string
        phase:
          enum: [lobby, answering, discussion, voting, round_result, finished]
        settings:
          $ref: "#/components/schemas/Settings"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        round:
          type: integer
        deadline:
          type: [string, "null"]
          format: date-time
        your_player_id:
          type: string
        your_prompt:
          type: string
        answer_locked:
          type: boolean
        vote_locked:
          type: boolean
        real_question:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        result:
          oneOf:
            - $ref: "#/components/schemas/RoundResult"
            - type: "null"
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
  responses:
    Problem:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"